refactor: 패키지 구조 변경 및 Mapper 레이어 삭제
