+++ src/main/java/cart/controller/AdminProductController.java
@@ -1,9 +1,8 @@
 package cart.controller;
 
-import cart.dto.ProductCreationRequest;
 import cart.dto.ProductDto;
-import cart.dto.ProductModificationRequest;
-import cart.mapper.ProductDtoMapper;
+import cart.dto.request.ProductCreationRequest;
+import cart.dto.request.ProductModificationRequest;
 import cart.service.ProductManagementService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -29,14 +28,14 @@ public class AdminProductController {
 
     @PostMapping
     public ResponseEntity<Void> postProducts(@Valid @RequestBody ProductCreationRequest request) {
-        final long id = managementService.save(ProductDtoMapper.from(request));
+        final long id = managementService.save(ProductCreationRequest.toProductDto(request));
         return ResponseEntity.created(URI.create("/products/" + id)).build();
     }
 
     @PutMapping("/{productId}")
     public ResponseEntity<Void> putProducts(@PathVariable Long productId,
                                             @Valid @RequestBody ProductModificationRequest request) {
-        managementService.update(ProductDtoMapper.from(request));
+        managementService.update(ProductModificationRequest.toProductDto(request));
         return ResponseEntity.noContent().build();
     }
