+++ src/main/java/cart/dao/JdbcMemberDao.java
@@ -1,7 +1,6 @@
 package cart.dao;
 
 import cart.entity.MemberEntity;
-import cart.entity.ProductEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
