+++ src/main/java/cart/dto/request/ProductModificationRequest.java
@@ -1,13 +1,18 @@
-package cart.dto;
+package cart.dto.request;
 
+import cart.dto.ProductDto;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.hibernate.validator.constraints.URL;
 
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
 
-public class ProductCreationRequest {
+public class ProductModificationRequest {
+
+    @NotNull
+    private final Long id;
 
     @NotBlank
     private final String name;
@@ -20,13 +25,24 @@ public class ProductCreationRequest {
     private final Integer price;
 
     @JsonCreator
-    public ProductCreationRequest(
+    public ProductModificationRequest(
+            @JsonProperty(value = "id") final Long id,
             @JsonProperty(value = "name") final String name,
             @JsonProperty(value = "image") final String image,
-            @JsonProperty(value = "price") final Integer price) {
+            @JsonProperty(value = "price") final Integer price
+    ) {
+        this.id = id;
         this.name = name;
         this.image = image;
         this.price = price;
+    }
+
+    public static ProductDto toProductDto(final ProductModificationRequest request) {
+        return ProductDto.of(request.getId(), request.getName(), request.getImage(), request.getPrice());
+    }
+
+    public Long getId() {
+        return id;
     }
 
     public String getName() {
