+++ src/main/java/cart/service/ProductManagementService.java
@@ -3,8 +3,6 @@ package cart.service;
 import cart.dao.ProductDao;
 import cart.dto.ProductDto;
 import cart.entity.ProductEntity;
-import cart.mapper.ProductDtoMapper;
-import cart.mapper.ProductEntityMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,23 +18,23 @@ public class ProductManagementService {
 
     public List<ProductDto> findAll() {
         final List<ProductEntity> productEntities = productDao.selectAll();
-        return ProductDtoMapper.from(productEntities);
+        return ProductDto.from(productEntities);
     }
 
     public long save(final ProductDto productDto) {
-        return productDao.insert(ProductEntityMapper.from(productDto));
+        return productDao.insert(ProductDto.toEntity(productDto));
     }
 
     public void update(final ProductDto productDto) {
-        int updatedRowCount = productDao.update(ProductEntityMapper.from(productDto));
-        if(updatedRowCount == 0){
+        int updatedRowCount = productDao.update(ProductDto.toEntity(productDto));
+        if (updatedRowCount == 0) {
             throw new IllegalArgumentException("존재하지 않는 상품입니다.");
         }
     }
 
     public void delete(final ProductDto productDto) {
-        int deletedRowCount = productDao.delete(ProductEntityMapper.from(productDto));
-        if(deletedRowCount == 0){
+        int deletedRowCount = productDao.delete(ProductDto.toEntity(productDto));
+        if (deletedRowCount == 0) {
             throw new IllegalArgumentException("존재하지 않는 상품입니다.");
         }
     }
